Make AudioParameters a class instead of a struct
Also collapses some long parameter lists into AudioParameters and moves some
of the hardcoded values (e.g. 16 bit audio in AudioDevice) to more appropriate
locations.

BUG=115902
TEST=manually testing everything works out

Review URL: https://chromiumcodereview.appspot.com/9655018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@128054 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: cfb09c2d58ecb16b14192619695ea54570344ca7

+++ content/renderer/pepper/pepper_plugin_delegate_impl.cc
@@ -598,8 +598,8 @@ PepperPluginDelegateImpl::CreateAudioOutput(
     uint32_t sample_rate,
     uint32_t sample_count,
     webkit::ppapi::PluginDelegate::PlatformAudioCommonClient* client) {
-  return PepperPlatformAudioOutputImpl::Create(sample_rate, sample_count,
-                                               client);
+  return PepperPlatformAudioOutputImpl::Create(
+      static_cast<int>(sample_rate), static_cast<int>(sample_count), client);
 }
 
 webkit::ppapi::PluginDelegate::PlatformAudioInput*
@@ -607,8 +607,8 @@ PepperPluginDelegateImpl::CreateAudioInput(
     uint32_t sample_rate,
     uint32_t sample_count,
     webkit::ppapi::PluginDelegate::PlatformAudioCommonClient* client) {
-  return PepperPlatformAudioInputImpl::Create(sample_rate, sample_count,
-                                              client);
+  return PepperPlatformAudioInputImpl::Create(
+      static_cast<int>(sample_rate), static_cast<int>(sample_count), client);
 }
 
 // If a broker has not already been created for this plugin, creates one.
